moved legacy fixes into the fixes module
