+++ src/rook/fixes/providers.py
@@ -41,14 +41,14 @@ class LegacyDatasetFixProvider(FixProvider):
     name = "legacy"
 
     def prepare(self, ds, *, context=None):
-        from rook.utils.decadal_fixes import decadal_fix_calendar
+        from rook.fixes.legacy_decadal import decadal_fix_calendar
 
         # TODO: decide whether this special CMIP6-decadal pre-concat calendar
         # preparation belongs in Woodpecker or remains a Rook operation hook.
         return decadal_fix_calendar(None, ds)
 
     def apply(self, ds, *, context=None):
-        from rook.utils.decadal_fixes import apply_decadal_fixes
+        from rook.fixes.legacy_decadal import apply_decadal_fixes
 
         context = context or FixContext()
         return apply_decadal_fixes(
@@ -64,7 +64,7 @@ class WoodpeckerDatasetFixProvider(FixProvider):
     name = "woodpecker"
 
     def prepare(self, ds, *, context=None):
-        from rook.utils.decadal_fixes import decadal_fix_calendar
+        from rook.fixes.legacy_decadal import decadal_fix_calendar
 
         # TODO: decide whether this special CMIP6-decadal pre-concat calendar
         # preparation belongs in Woodpecker or remains a Rook operation hook.
